// src/classes/Entity/types.d.ts
export declare type KeyAttributeDefinition = {
    type: 'string' | 'number' | 'bigint' | 'binary';
    default: any;
    hidden: boolean;
    delimiter: string;
    prefix: string;
    suffix: string;
    onUpdate: boolean;
    dependsOn: string | string[];
    transform: (value: any, data: any) => any;
    format: (value: any, data: any) => any;
    coerce: boolean;
    save: never;
    required: never;
    alias: never;
    map: never;
    setType: never;
};
export declare type PartitionKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey: true;
    sortKey?: false;
};
export declare type GSIPartitionKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey: string;
    sortKey?: false;
};
export declare type SortKeyDefinition = Partial<KeyAttributeDefinition> & {
    sortKey: true;
    partitionKey?: false;
};
export declare type GSISortKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey?: false;
    sortKey: string;
};
export declare type PureAttributeDefinition = Partial<{
    partitionKey: false;
    sortKey: false;
    type: DynamoDBTypes;
    default: any | ((data: object) => any);
    dependsOn: string | string[];
    transform: (value: any, data: {}) => any;
    format: (value: any, data: {}) => any;
    coerce: boolean;
    save: boolean;
    onUpdate: boolean;
    hidden: boolean;
    required: boolean | 'always';
    alias: string;
    map: string;
    setType: DynamoDBKeyTypes;
    delimiter: string;
    prefix: string;
    suffix: string;
}>;